refector(Horse): 마 기물의 경로 구하기 인터페이스로 구현
